ADD vGus

+++ MCU_Application/MainApp/vgus.h
@@ -6,8 +6,8 @@
 /*****************************************************************************/
 void vGusRegRead(uint8_t addr, uint8_t length);//读取VGUS寄存器
 void vGusRegWrite(uint8_t addr, uint8_t *pbuf ,uint8_t length);//写入VGUS寄存器
-void vGusVarRead(uint8_t addr, uint8_t length);//读取VGUS变量
-void vGusVarWrite(uint8_t addr, uint8_t *pbuf ,uint8_t length);//写入VGUS变量
+void vGusVarRead(uint16_t addr, uint8_t length);//读取VGUS变量
+void vGusVarWrite(uint16_t addr, uint16_t *pbuf ,uint8_t length);//写入VGUS变量
 /*****************************************************************************/
 
 #endif
